// components/brave_wallet_ui/constants/types.ts
export type SafeBlowfishWarning = {
  severity: BraveWallet.BlowfishWarningSeverity
  kind: BraveWallet.BlowfishWarningKind
  /**
   * human-readable message to present to the end-user
   */
  message: string
}
// ...
interface SafeBlowfishEVMError {
  kind: BraveWallet.BlowfishEVMErrorKind
  humanReadableError: string
}

interface SafeBlowfishSolanaError {
  kind: BraveWallet.BlowfishSolanaErrorKind
  humanReadableError: string
}
// ...
export type SafeEvmEvent = SafeEvmApprovalEvent | SafeEvmTransferEvent
// ...
type SafeSolanaEvent =
  | SafeSolTransferEvent
  | SafeSplApprovalEvent
  | SafeSplTransferEvent
  | SafeSolanaStakeChangeEvent
  | SafeSolanaAccountOwnerChangeEvent

export interface SafeEVMSimulationResults {
  error: SafeBlowfishEVMError | undefined
  expectedStateChanges: SafeEvmEvent[]
};

export interface SafeSolanaSimulationResults {
  error: SafeBlowfishSolanaError | undefined
  expectedStateChanges: SafeSolanaEvent[]
};
// ...
type SafeBlowfishResponseBase = {
  action: BraveWallet.BlowfishSuggestedAction
  warnings: SafeBlowfishWarning[]
}

export type SafeBlowfishEvmResponse = SafeBlowfishResponseBase & {
  simulationResults: SafeEVMSimulationResults
}

export type SafeBlowfishSolanaResponse = SafeBlowfishResponseBase & {
  simulationResults: SafeSolanaSimulationResults
}